+++ src/main/java/thedarkcolour/exdeorum/event/ClientsideCode.java
@@ -20,7 +20,6 @@ package thedarkcolour.exdeorum.event;
 
 import net.minecraft.client.Minecraft;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.level.Level;
 
 // necessary to avoid EventBus loading LocalPlayer through its ASM transformations
 class ClientsideCode {